refactor(frontend): refactor to use reusable component
