+++ apps/iaimc-frontend/src/app/customer/landing-page/landing-page.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { StageComponent } from '@igus-accelerator-injection-molding-configurator/ui-components';
 import { ModelUploadComponent } from '../../../../../../libs/ui-components/src/model/components/model-upload/model-upload.component';
-import { FooterComponent } from '../../../../../../libs/ui-components/src/landing-page/components/footer/footer.component';
+import { FooterComponent } from '../../../../../../libs/ui-components/src/landing-page/footer/footer.component';
 
 @Component({
   selector: 'app-landing-page',
