--- libs/ui-components/src/landing-page/components/footer/footer.component.html
@@ -1,101 +1,0 @@
-<mat-toolbar>
-  <footer>
-    <form>
-      <section class="model-container">
-        <div class="model-container__logo">
-          <app-model-logo />
-          <p>{{ 'footer.LOGO_TEXT' | transloco }}</p>
-        </div>
-        <p>
-          {{ 'footer.MODEL_DESCRIPTIO' | transloco }}
-        </p>
-      </section>
-      <section class="subscription-container">
-        <h4>{{ 'footer.SUBSCRIPTION_HEADER' | transloco }}</h4>
-        <div class="form-fields">
-          <div>
-            <mat-form-field class="small">
-              <input
-                type="text"
-                matInput
-                [formControl]="getField('firstName')"
-                [placeholder]="'footer.FIRST_NAME' | transloco"
-              />
-            </mat-form-field>
-            @if (
-              subscriptionForm.get('firstName')?.hasError('required') &&
-              subscriptionForm.get('firstName')?.touched
-            ) {
-              <mat-error>{{ 'footer.REQUIRED_FIELD' | transloco }}</mat-error>
-            } @else if (
-              subscriptionForm.get('firstName')?.hasError('pattern')
-            ) {
-              <mat-error>{{ 'footer.ENTER_VALID_NAME' | transloco }}</mat-error>
-            } @else if (
-              subscriptionForm.get('firstName')?.hasError('minlength')
-            ) {
-              <mat-error>{{ 'footer.NAME_TOO_SHORT' | transloco }}</mat-error>
-            } @else if (
-              subscriptionForm.get('firstName')?.hasError('maxlength')
-            ) {
-              <mat-error>{{ 'footer.NAME_TOO_LONG' | transloco }}</mat-error>
-            }
-          </div>
-
-          <div>
-            <mat-form-field class="small">
-              <input
-                matInput
-                [formControl]="getField('email')"
-                [placeholder]="'footer.EMAIL' | transloco"
-              />
-            </mat-form-field>
-            @if (subscriptionForm.get('email')?.hasError('email')) {
-              <mat-error>{{
-                'footer.ENTER_VALID_EMAIL' | transloco
-              }}</mat-error>
-            }
-            @if (
-              subscriptionForm.get('email')?.hasError('required') &&
-              subscriptionForm.get('email')?.touched
-            ) {
-              <mat-error>{{ 'footer.REQUIRED_FIELD' | transloco }}</mat-error>
-            }
-          </div>
-          <button class="custom-button" mat-button>
-            <span> {{ 'footer.SUBSCRIBE_TEXT' | transloco }}</span>
-            @if (isSubscriptionLoading()) {
-              <mat-spinner [diameter]="20" [strokeWidth]="2"></mat-spinner>
-            }
-          </button>
-        </div>
-      </section>
-    </form>
-    <section class="footer-links">
-      <div class="footer-links__terms-and-condition">
-        <p>© {{ currentYear }} IMDesigner</p>
-        <a routerLink="#">Terms of Service</a>
-        <a routerLink="#">Privacy & Cookies policy</a>
-      </div>
-      <div class="icons-group">
-        <a routerLink="#">
-          <button mat-icon-button class="icons-group__button">
-            <mat-icon><img src="assets/images/github.svg" alt="" /></mat-icon>
-          </button>
-        </a>
-        <a routerLink="#">
-          <button mat-icon-button class="icons-group__button">
-            <mat-icon><img src="assets/images/x.svg" alt="" /></mat-icon>
-          </button>
-        </a>
-        <a routerLink="#">
-          <button mat-icon-button class="icons-group__button">
-            <mat-icon
-              ><img src="assets/images/linked-in.svg" alt=""
-            /></mat-icon>
-          </button>
-        </a>
-      </div>
-    </section>
-  </footer>
-</mat-toolbar>
