+++ libs/ui-components/src/landing-page/footer/footer.component.ts
@@ -15,9 +15,11 @@ import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
 import { RouterLink } from '@angular/router';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { ModelLogoComponent } from '../../../svgs/model-logo/model-logo.component';
-import { TranslocoPipe } from '@jsverse/transloco';
-import { formField } from '../../../utilities/helper-function';
+import { ModelLogoComponent } from '../../svgs/model-logo/model-logo.component';
+import { TranslocoPipe, TranslocoService } from '@jsverse/transloco';
+import { formField } from '../../utilities/helper-function';
+import { ReusableFormFieldComponent } from "../../reusable-components/reusable-form-field/reusable-form-field.component";
+import { ReusableButtonComponent } from '../../reusable-components/reusable-button/reusable-button.component';
 
 @Component({
   selector: 'app-footer',
@@ -31,14 +33,17 @@ import { formField } from '../../../utilities/helper-function';
     MatInputModule,
     RouterLink,
     MatIconModule,
-  ],
+    ReusableFormFieldComponent,
+    ReusableButtonComponent
+],
   templateUrl: './footer.component.html',
   styleUrl: './footer.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class FooterComponent {
   private readonly fb = inject(FormBuilder);
-  public currentYear = new Date().getFullYear();
+  private readonly translocoService = inject(TranslocoService);
+  public readonly currentYear = new Date().getFullYear();
   public readonly isSubscriptionLoading = signal<boolean>(false);
 
   public subscriptionForm = this.fb.group({
@@ -57,4 +62,12 @@ export class FooterComponent {
   public getField(field: string) {
     return formField(field, this.subscriptionForm);
   }
+
+  public readonly errorMessages = {
+    required: this.translocoService.translate('footer.REQUIRED_FIELD'),
+    pattern: this.translocoService.translate('footer.ENTER_VALID_NAME'),
+    email: this.translocoService.translate('footer.ENTER_VALID_EMAIL'),
+    minlength: this.translocoService.translate('footer.NAME_TOO_SHORT'),
+    maxlength: this.translocoService.translate('footer.NAME_TOO_LONG')
+  };
 }
