+++ libs/ui-components/src/landing-page/footer/footer.i18n.json
@@ -11,7 +11,7 @@
       "ru": "Дизайнер моделей для литья под давлением",
       "zh": "注射模型设计师"
     },
-    "MODEL_DESCRIPTIO": {
+    "MODEL_DESCRIPTION": {
       "ar": "حوّل نماذجك ثلاثية الأبعاد إلى واقع مع خدمات القولبة بالحقن الاحترافية",
       "de": "Verwandeln Sie Ihre 3D-Modelle mit professionellen Spritzgussdiensten in Realität",
       "en": "Transform your 3D models into reality with professional injection molding services",
