+++ libs/ui-components/src/reusable-components/reusable-button/reusable-button.component.html
@@ -24,6 +24,7 @@
   <button
     mat-button
     [class]="className"
+    [class.footer]="isInFooter"
     [type]="type"
     [disabled]="disabled"
     [color]="color"
