+++ libs/ui-components/src/reusable-components/reusable-form-field/reusable-form-field.component.html
@@ -11,7 +11,7 @@
   } @else if (isTextarea) {
     <textarea matInput [formControl]="control"></textarea>
   } @else {
-    <input matInput [formControl]="control" [type]="inputType" />
+    <input [class.footer]="isInFooter" matInput [formControl]="control" [type]="inputType" />
   }
 
   @if (control.touched) {
